Refactor code structure and add setup.py file
Upgraded the structure of the project for better organization, modularity, and easier deployment. The main.py and utils.py files were removed and their functionality was divided among new modules located inside the src/ directory. The file structure was changed for .gitignore and README.md to provide more specific and necessary information. A setup.py file was created to handle the package requirement for easier setup and deployment of the project. Code in validations.py was updated to handle OS directory format for better OS compatibility.