Add openpyxl dependency and optimize ExcelReport module
Updated the requirements.txt to include openpyxl package which is needed for Excel operations. In ExcelReport.py, removed the abc module as it was not being used and made changes to improve the readability and efficiency of the code. The function '_save' was removed as 'close' already performs save operation. Also added os.startfile to auto open the created file.