Update README with setup.py execution info
Added the instruction about the use of setup.py in the README.md file. This change provides clarity that there is no need to install the libraries one by one. Simultaneously, it guides the users on how to use the setup.py file to install all the dependencies at once. This aids in hassle-free project setup and improves user experience.