Improve user prompts with colorama package
The commit introduces colorama in user prompts to improve user experience. Fore and Style are added and implemented in inputs and print statements for improved visibility and readability. Additionally, a more robust validation is added for Windows, Mac, and Linux file paths in the ValidationData module, increasing the tool's cross-platform compatibility.

Signed-off-by: tihamerkarlovich.aldana.contractor <tihamerkarlovich.aldana.contractor@bbva.com>