Update file path handling and input validation
Several changes have been made to improve the handling of file path in the code and enhance input validation. In README.md, Blu.js has been renamed to BluU.js and the 'abc' directory is removed. In main.py, Better formatting on user prompt is added to support either Bitbucket or GitHub. In DataGenerator.py, the 'URL rama' field is moved to the end of the dictionary. Lastly, in ValidationData.py, the input validation now includes 'D' and 'R090' – followed by particular spaces – along with 'M' and 'A', as valid starts for input data format, to better accord with diff format specification.

Signed-off-by: Tihamer Karlovich Aldana <tihameraldana@icloud.com>